requirements.txt
bson==0.5.8
Flask==1.1.1
flask_restful==0.3.7
flask_script==2.0.6
flask_migrate==2.5.2
flask_sqlalchemy==2.4.0
marshmallow-sqlalchemy==0.19.0
requests==2.22.0
pymongo==3.9.0
flask_pymongo==2.3.0
Flask_JWT_Extended==3.22.0
Flask_Bcrypt==0.7.1
jsonschema==3.0.2
python-dotenv==0.10.3
pandas
flask_swagger_ui==3.20.9
flask-restplus
pyfiglet
google-cloud-storage
dnspython
flask-cors
numpy
psycopg2
flask_restplus
flask_marshmallow
tpot
sklearn
scipy
scikit-learn
joblib
deap
update_checker
tqdm
stopit
matplotlib
flask_jsonpify
seaborn
graphviz
pydotplus
Pillow
marshmallow
fbprophet